+++ src/__generated__/Artwork_artwork.graphql.ts
@@ -6,6 +6,7 @@ import { AboutWork_artwork$ref } from "./AboutWork_artwork.graphql";
 import { ArtworkAvailability_artwork$ref } from "./ArtworkAvailability_artwork.graphql";
 import { ArtworkDetails_artwork$ref } from "./ArtworkDetails_artwork.graphql";
 import { ArtworkHeader_artwork$ref } from "./ArtworkHeader_artwork.graphql";
+import { ArtworkHistory_artwork$ref } from "./ArtworkHistory_artwork.graphql";
 import { OtherWorks_artwork$ref } from "./OtherWorks_artwork.graphql";
 import { PartnerCard_artwork$ref } from "./PartnerCard_artwork.graphql";
 import { SellerInfo_artwork$ref } from "./SellerInfo_artwork.graphql";
@@ -15,6 +16,9 @@ export type Artwork_artwork = {
     readonly availability: string | null;
     readonly additional_information: string | null;
     readonly description: string | null;
+    readonly provenance: string | null;
+    readonly exhibition_history: string | null;
+    readonly literature: string | null;
     readonly layer: {
         readonly artworksConnection: {
             readonly edges: ReadonlyArray<{
@@ -47,7 +51,7 @@ export type Artwork_artwork = {
             } | null> | null;
         } | null;
     } | null;
-    readonly " $fragmentRefs": ArtworkAvailability_artwork$ref & PartnerCard_artwork$ref & SellerInfo_artwork$ref & AboutWork_artwork$ref & OtherWorks_artwork$ref & AboutArtist_artwork$ref & ArtworkDetails_artwork$ref & ArtworkHeader_artwork$ref;
+    readonly " $fragmentRefs": ArtworkAvailability_artwork$ref & PartnerCard_artwork$ref & SellerInfo_artwork$ref & AboutWork_artwork$ref & OtherWorks_artwork$ref & AboutArtist_artwork$ref & ArtworkDetails_artwork$ref & ArtworkHeader_artwork$ref & ArtworkHistory_artwork$ref;
     readonly " $refType": Artwork_artwork$ref;
 };
 
@@ -138,6 +142,27 @@ return {
       "kind": "ScalarField",
       "alias": null,
       "name": "description",
+      "args": null,
+      "storageKey": null
+    },
+    {
+      "kind": "ScalarField",
+      "alias": null,
+      "name": "provenance",
+      "args": null,
+      "storageKey": null
+    },
+    {
+      "kind": "ScalarField",
+      "alias": null,
+      "name": "exhibition_history",
+      "args": null,
+      "storageKey": null
+    },
+    {
+      "kind": "ScalarField",
+      "alias": null,
+      "name": "literature",
       "args": null,
       "storageKey": null
     },
@@ -284,9 +309,14 @@ return {
       "kind": "FragmentSpread",
       "name": "ArtworkHeader_artwork",
       "args": null
+    },
+    {
+      "kind": "FragmentSpread",
+      "name": "ArtworkHistory_artwork",
+      "args": null
     }
   ]
 };
 })();
-(node as any).hash = '2c496d38547b9a35719a7edbd6ec884b';
+(node as any).hash = '38aa50c794e853557cd263c80aed2864';
 export default node;