+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Use gravity ID when opening artwork from artwork grid instead of global object ID - alloy
 - Fixes a React Native font-rendering issue - david & ash
 - Adds artwork classifications page and links to it from artwork page - lilyfromseattle
+- Add artwork history sections - steve
 
 ### 1.12.1